Move Spreadsheet.ts to the root of src.
