--- functions/src/GoogleSheets/index.ts
@@ -1,1 +1,0 @@
-export * from './Spreadsheet';
